数据库系统课程设计/project/src/requirements.txt
Flask
psycopg2-binary
Flask-Login 